Level2

+++ Project 0/Assets/Script/Bomb/SO/BombSO.cs
@@ -5,7 +5,7 @@ namespace Wepons.Bomb
     [CreateAssetMenu(fileName = "Bomb", menuName = "ScriptableObjects/bombSO")]
     public class BombSO : ScriptableObject
     {
-        public float bombDamage;
+        public int bombDamage;
         public float DamageRadius;
         public float primingTime;
     }
